Benchmarks overhaul: - Add all POSIX versions - Add all fixed versions - All currently available infomation for each one in `info.yaml` (including ground truth)

+++ benchmarks/high_profile/c00-steam/fixed.sh
@@ -356,7 +356,7 @@ reset_steam()
 	done
 
 	# Scary!
-	rm -rf "$STEAMROOT/"* # bug here: will delete root if line 17 fails
+	rm -rf "${STEAMROOT:?}/"*
 
 	# Move things back into place
 	mv -f "$STEAM_SAVE/"* "$STEAMROOT/"
